requirements/last.txt
amqp>=5.3.1
annotated-types>=0.7.0
anyascii>=0.3.2
anyio>=4.8.0
argon2-cffi>=23.1.0
argon2-cffi-bindings>=21.2.0
asgiref>=3.8.1
async-timeout>=5.0.1
attrs>=25.1.0
beautifulsoup4>=4.12.3
billiard>=4.2.1
celery>=5.4.0
certifi>=2025.1.31
cffi>=1.17.1
charset-normalizer>=3.4.1
click>=8.1.8
click-didyoumean>=0.3.1
click-plugins>=1.1.1
click-repl>=0.3.0
colorama>=0.4.6
crispy-bootstrap5>=2024.2
cron-descriptor>=1.4.5
cryptography>=44.0.1
defusedxml>=0.7.1
distro>=1.9.0
Django>=5.0.8
django-allauth>=64.0.0
django-anymail>=11.1
django-celery-beat>=2.6.0
django-cors-headers>=4.4.0
django-crispy-forms>=2.3
django-environ>=0.11.2
django-filter>=24.3
django-model-utils>=4.5.1
django-modelcluster>=6.4
django-permissionedforms>=0.1
django-phonenumber-field>=8.0.0
django-redis>=5.4.0
django-stubs-ext>=5.1.3
django-taggit>=6.1.0
django-tasks>=0.6.1
django-timezone-field>=7.1
django-treebeard>=4.7.1
djangorestframework>=3.15.2
draftjs-exporter>=5.0.0
drf-spectacular>=0.27.2
et_xmlfile>=2.0.0
fido2>=1.2.0
filetype>=1.2.0
flower>=2.0.1
gunicorn>=23.0.0
h11>=0.14.0
hiredis>=3.0.0
httpcore>=1.0.7
httpx>=0.28.1
humanize>=4.11.0
idna>=3.10
inflection>=0.5.1
jiter>=0.8.2
jsonschema>=4.23.0
jsonschema-specifications>=2024.10.1
kombu>=5.4.2
laces>=0.1.2
openai>=1.63.0
openpyxl>=3.1.5
packaging>=24.2
phonenumberslite>=8.13.54
pillow>=10.4.0
pillow_heif>=0.21.0
prometheus_client>=0.21.1
prompt_toolkit>=3.0.50
psycopg>=3.2.1
psycopg-c>=3.2.1
pycparser>=2.22
pydantic>=2.10.6
pydantic_core>=2.27.2
python-crontab>=3.2.0
python-dateutil>=2.9.0.post0
python-slugify>=8.0.4
pytz>=2025.1
PyYAML>=6.0.2
qrcode>=8.0
redis>=5.0.8
referencing>=0.36.2
requests>=2.32.3
rpds-py>=0.22.3
six>=1.17.0
sniffio>=1.3.1
soupsieve>=2.6
sqlparse>=0.5.3
telepath>=0.3.1
text-unidecode>=1.3
tornado>=6.4.2
tqdm>=4.67.1
typing_extensions>=4.12.2
tzdata>=2025.1
uritemplate>=4.1.1
urllib3>=2.3.0
vine>=5.1.0
wagtail>=6.4
wagtailcodeblock>=1.29.0.2
wcwidth>=0.2.13
whitenoise>=6.7.0
Willow>=1.9.0
wagtail-seo
python-magic